Assign hostname to 127.0.1.1 if available
This patch assigns `@machine.config.vm.hostname` to 127.0.1.1 and falls back to
`@machine.name` if the hostname was not set.

The hostname can be set to something resembling a fqdn, which allows `facter
fqdn` to return a value. This in turn prevents a bunch of problems that can
occur in the Puppet ecosystem if Facter cannot determine a fully qualified
domain name.