Adding support for sync'ing local hosts.
Adds two commands to the plugin which can be used to manage the hosts file on the local machine. The current implemention requires that the vagrant boxes are running. In the future it might be nice to do something smarter with configuration objects. But this is a good start and solves a lot of pain.