Provision Windows using an elevated shell
For the windows communicator, `sudo` is aliased to `execute`. There is a
special `elevated` option that can be invoked. This patch sets `elevated` to
true when provision hosts.

The reasoning behind this actually has nothing to do with permissions. The
elevated mode of WinRM execution actually uploads the shell command as a
temporary script file instead of executing it directly as a string. This is
important for Windows as setting too many host definitions in a large command
line string will result in a "The command line is too long" error. A more
elegant solution to this problem may be to hand off to the actual shell
provisioner and let it take care of the gory details of command execution.

For communicators that do not support the `elevated` option, such as SSH and
the old vagrant-windows plugin, the `elevated` argument should be silently
ignored.

Fixes #37.