Support fqdn-named Windows Vagrant machines
Windows machine names cannot contain special characters and cannot be
longer than 15 characters in length. However, it's reasonable to define
Vagrant machines named by FQDN rather than windows-compliant short name.

This commit modifies the Windows change_hostname method to sanitize the
vagrant machine name before attempting to change the name in the guest
OS. It does so by converting illegal characters to "-", and if the name
is too long truncating at the first illegal character.

The effective result of this should be that if a vagrant Windows machine
has a name such as "win1.longdomain.evenlonger.net", the vagrant-hosts
provisioner will assign the guest OS the machine name "win1".