Add capability to sync hosts across VMs
Prior to this commit, a VM would only receive host entries from other VMs
that were already up and running in the environment. This means that for two
machines booted in order, A and B, A would have only its own entries while
B would have entries for both machines. A `vagrant provision A` would be
required to bring the environment up to parity.

This patch adds a new provisioner option, `sync_hosts`, that triggers an update
on all running machines if set to true. An update is also triggered on machine
destruction.

Closes oscar-stack/vagrant-hosts#17