Use all_machines helper to list active machines
Sometimes Vagrant will think a machine is active when it is actually gone.
Attempting to operate on this machine will result in a `MachineNotFound`
error.

This patch switches the `sync_hosts` implementation over to using the
`all_machines` helper in the `VagrantHosts::Addresses` module which contains
a check to guard against this issue.

Fixes oscar-stack/vagrant-hosts#53