Stop combining host entries based on IP
This commit stops the combination of hosts entries based on IP address. This
initially seemed like a good idea, but resulted in a lot of duplicate code
and made it difficult to determine which Vagrant configuration bits were
responsible for which host entries. This change also preserves the existing
behavior and is more suitable for a .z release.