Fix get_default_iface fact on Windows
This commit updates the network_facts capability so that we no longer chomp
newlines on Windows. This is required as Vagrant 1.9 and later moved to version
2.x of the WinRM gem which implements PSRP instead of running things through
CMD. PSRP returns each line separately instead of one big stdout block, so
chomping removes the ability to iterate by line.

In general, the chomp operation smells bad and should be dropped along with
Vagrant 1.8 and older.