Refactor Windows sync_hosts to match POSIX
This commit refactors the Windows implementation of the sync_hosts capability
to match the behavior of the posix implementation. Instead of uploading and
running a PowerShell script that iterates through a list of hostname pairs
and adding missing content, the capability now generates a hosts file locally,
uploads it, and copies it into place. This allows hostnames to be removed
from the Windows hosts file and fixes some bugs around length limits of
argument lists.

Fixes oscar-stack/vagrant-hosts#37.
Fixes oscar-stack/vagrant-hosts#51.